+++ crates/action-graph/tests/utils.rs
@@ -1,6 +1,7 @@
 use moon_action_graph::ActionGraphBuilder;
 use moon_platform::PlatformManager;
 use moon_project_graph::ProjectGraph;
+use moon_task_graph::TaskGraph;
 use moon_test_utils2::{
     generate_platform_manager_from_sandbox, generate_project_graph_from_sandbox,
 };
@@ -9,6 +10,7 @@ use std::path::Path;
 pub struct ActionGraphContainer {
     pub platform_manager: PlatformManager,
     pub project_graph: ProjectGraph,
+    pub task_graph: TaskGraph,
 }
 
 impl ActionGraphContainer {
@@ -16,10 +18,16 @@ impl ActionGraphContainer {
         Self {
             platform_manager: generate_platform_manager_from_sandbox(root).await,
             project_graph: generate_project_graph_from_sandbox(root).await,
+            task_graph: TaskGraph::default(), // TODO
         }
     }
 
     pub fn create_builder(&self) -> ActionGraphBuilder {
-        ActionGraphBuilder::with_platforms(&self.platform_manager, &self.project_graph).unwrap()
+        ActionGraphBuilder::with_platforms(
+            &self.platform_manager,
+            &self.project_graph,
+            &self.task_graph,
+        )
+        .unwrap()
     }
 }